adding click interface to python compiler

+++ src/zoombuild/tools/python_packager.py
@@ -5,17 +5,40 @@ import sys
 import zipfile
 from datetime import datetime
 from io import StringIO
+import logging
 
-__version__ = (0, 0, 1)
+from .binary_packager import METADATA_FILE
+from .project_info import PyProject
+import click
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+handler = logging.StreamHandler()
+formatter = logging.Formatter("{message}", style="{")
+handler.setFormatter(formatter)
+logger.addHandler(handler)
+
+
+__version__ = (0, 0, 2)
 __builder__ = __name__
-METADATA_FILE = "environment.ini"
 
 
-def default_filter(pth):
+def _default_filter(pth):
     return os.path.isdir(pth) or pth.endswith(".py")
 
 
-def compile_tree(source_tree, zipname, optimize=1, filter=default_filter):
+def compile_tree(project:PyProject, zipname, optimize=1, filter=_default_filter):
+    
+    logger.info("Compiling %s", project)
+    if not zipname:
+        zipname = os.path.join(project.project_root, project.name + ".zip")
+
+    source_tree = project.find_package_dir()
+    logger.info("Compiling %s", source_tree)
+
+    if not os.path.exists(source_tree):
+        raise RuntimeError("Could not find source directory in {project.name}")
+    
     delete_caches(source_tree)
 
     with zipfile.PyZipFile(zipname, "w", optimize=optimize) as archive:
@@ -38,7 +61,7 @@ def write_metadata(archive):
         "machine": os.getenv("COMPUTERNAME", "unknown"),
         "user": os.getenv("USERNAME", "unknown"),
         "version": __version__,
-        "tool": __builder__,
+        "tool": __name__,
     }
 
     tmp = StringIO()
@@ -67,9 +90,22 @@ def find_resource_files(source_tree):
             yield full_file, relpath
 
 
-if __name__ == "__main__":
-    if os.path.exists("test.zip"):
-        os.remove("test.zip")
 
-    # todo: proper CLI
-    compile_tree("src", "test.zip", optimize=1)
+@click.command(help='Package Python project into a distributable zip file')
+@click.argument('project')
+@click.option('--output',
+              default = None,
+              help = "Output path for the zip file"
+              )
+@click.option("--verbose", is_flag=True, help="Increase logging verbosity")
+@click.option('--optimize',
+              default = 1,
+              help = 'Optimize compiled output (default = 1, strips asserts and __DEBUG__)')
+def main(project, output, optimize, verbose):
+    if verbose:
+        logger.setLevel(logging.DEBUG)
+    if not os.path.exists(project):
+        raise ValueError("Project {project} not found")
+
+    prj = PyProject(project)
+    compile_tree(prj, output, optimize)
